[core] Code refactoring
